Move colors to ui folder
