+++ autothumb.py
@@ -23,9 +23,10 @@ import tempfile
 
 import bpy
 
-from . import bg_blender, colors, paths, utils
+from . import bg_blender, paths, utils
 from .config import HANA3D_DESCRIPTION, HANA3D_NAME
 from .report_tools import execute_wrapper
+from .src.ui import colors
 from .src.ui.main import UI
 
 HANA3D_EXPORT_DATA_FILE = f"{HANA3D_NAME}_data.json"
