+++ download.py
@@ -44,9 +44,10 @@ from .config import (
 from .report_tools import execute_wrapper
 from .src.search.query import Query
 from .src.search.search import Search
+from .src.ui import colors
 from .src.ui.main import UI
 
-from . import append_link, colors, hana3d_types, logger, paths, render_tools, utils  # noqa E501 isort:skip
+from . import append_link, hana3d_types, logger, paths, render_tools, utils  # noqa E501 isort:skip
 
 
 download_threads = {}
