+++ ui.py
@@ -28,12 +28,12 @@ from bpy.props import BoolProperty, StringProperty
 from bpy_extras import view3d_utils
 from mathutils import Vector
 
-from . import bg_blender, colors, download, paths, render, search, utils
+from . import bg_blender, download, paths, render, search, utils
 from .config import HANA3D_DESCRIPTION, HANA3D_MODELS, HANA3D_NAME, HANA3D_UI
 from .report_tools import execute_wrapper
 from .src.preferences.preferences import Preferences
 from .src.search.search import Search
-from .src.ui import bgl_helper
+from .src.ui import bgl_helper, colors
 from .src.ui.main import UI
 
 handler_2d = None
